Se modifica modelo para aceptar decimales, se agrega traduccion de datatables , se edita layou para mostrar solo la barra de navegacion cuando se ha logeado, se realiza scaffolding de Identiy

+++ src/MarcosM3.Web/Models/CotizacionViewModel.cs
@@ -8,9 +8,9 @@ namespace MarcosM3.Web.Models
     {
         public int Cantidad { get; set; }
 
-        public int Largo { get; set; }
+        public decimal Largo { get; set; }
 
-        public int Ancho { get; set; }
+        public decimal Ancho { get; set; }
 
         public Moldura Moldura { get; set; }
